Biblioteca/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asgiref==3.7.2
attrs==24.2.0
blinker==1.9.0
bottle==0.13.4
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
cryptography==45.0.7
discord.py==2.4.0
Django==4.2.2
django-cors-headers==4.3.1
Flask==3.1.2
Flask-Cors==4.0.1
Flask-MySQL==1.5.2
frozenlist==1.5.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.17.0
multidict==6.1.0
oracledb==3.3.0
packaging==24.0
propcache==0.2.1
proxy_tools==0.1.0
pycparser==2.23
PyMySQL==1.1.0
python-dotenv==1.1.1
pythonnet==3.0.5
pywebview==5.3.2
requests==2.31.0
sqlparse==0.4.4
transbank-sdk==5.0.0
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.2.1
Werkzeug==3.1.3
whitenoise==6.6.0
yarl==1.18.3